requirements.txt
flask>=3.1.1
folium>=0.20.0
reportlab>=4.0.0
Pillow>=10.0.0
werkzeug>=3.0.0
typing-extensions>=4.0.0
